commited some hangling uncommits

+++ include/uvector.h
@@ -50,7 +50,8 @@ public:
     typedef std::reverse_iterator<const_iterator>	const_reverse_iterator;
     typedef std::reverse_iterator<iterator>		reverse_iterator;
 
-    typedef size_t					size_type;
+    // typedef size_t					size_type;
+    typedef uint					size_type;
     typedef ptrdiff_t					difference_type;
 
 public:
